Adding distance and overlap calculations
Distance is the length of the path of a drive.

Overlap is a segment of a drive which corresponds with a segment of
another drive within a certain [time, distance] error.

Runtime is 42 sec for reading all drives and analyzing 500 of them (so
about half an hour for whole set).  This is with time tolerance of 5 min;
would presumably be higher for larger time tolerances.